Divide HUGE Player.h/cpp to separate files
